API polish: one Rect, nine-slices, text options, float32 input, buses, embedded assets
lin gains Rect (used by clips, the UI, cameras and nine-slices), Vec2
and Vec3 helpers, Euler angles, QuatLookAt, Mat3, Decompose. gfx.Color
gains Lerp, Mul, Scale, HSV and Premultiplied. Text drawing collapses to
DrawText and DrawTextBlock with Size, Angle and Baseline in TextOptions
for every font kind, and Font.Measure takes the options; fonts expose
Descent. Camera gains an orthographic mode and Graphics.Project maps
world points to the view. gfx.NineSlice and gfx.Region are values;
ui.Rect and ui.Slice are the gfx types. Input positions are float32 and
KeyRepeated reports OS repeats. ecs gains Each3 and Each4.

audio: buses (music, effects, dialogue and NewBus) with ramped volume
and pause, Mixer.SetPaused, Voice.Position, Seek and OnDone,
Sound.Duration, Music.Duration and Seek, byte-taking decoders.
asset: OpenFS with Dir, PackFile and FSSource so go:embed works, and
one-call Image, Texture, Font, SDFFont, Sound, Music, Model and Tracker
loaders.

+++ examples/inputs/main.go
@@ -24,9 +24,9 @@ type game struct {
 
 	font     *gfx.Font
 	typed    []rune
-	scroll   float64
-	deltaX   float64
-	deltaY   float64
+	scroll   float32
+	deltaX   float32
+	deltaY   float32
 	shotDone bool
 }
